requirements.txt
requests>=2.31.0
websocket-client>=1.6.0
pyyaml>=6.0
python-dotenv>=1.0.0
